Add cross-publishing sync workflow and publish 6 articles
- Add sync-writing script: syncs Obsidian newsletters marked
  website:true to src/content/writing/ with Substack content
  cleaning, tag display transforms, and idempotent output
- Add normalize-obsidian-newsletters script: brings all 194
  newsletter files to canonical schema (title, tags, field order)
- Remove migrate-obsidian-newsletters script (one-time, complete)
- Add canonicalUrl to writing schema and WritingLayout for SEO
- Show description as article deck in WritingLayout
- Fix date timezone offset on article pages, writing index, homepage
- Publish 6 articles: AI Runs on Text, Voice is the Missing Link,
  Curate Your Context, Your Instructions Now Run Like Code,
  Prototype Code Is Not Production Code, Designers are a Rare Breed

Closes #21
