+++ src/content/writing/designers-are-a-rare-breed.md
@@ -1,0 +1,53 @@
+---
+title: Designers are a Rare Breed
+description: AI is raising the floor. The ceiling is yours.
+publishedDate: 2026-05-10
+categories:
+  - Newsletter
+tags:
+  - AI
+  - Design
+image: /images/writing/img-designers-rare-breed.jpg
+canonicalUrl: "https://www.unknownarts.co/p/designers-are-a-rare-breed"
+draft: false
+---
+Designers are a rare breed.
+
+It’s easy to forget if you’ve built your community around design, but the data is clear: [designers make up roughly 0.25% of the US workforce](https://www.bls.gov/ooh/arts-and-design/graphic-designers.htm), or about one in every 400 workers. [Only 16% of US companies](https://ncses.nsf.gov/pubs/nsf25330/assets/nsf25330.pdf) say they pursue design as a structured process. The rest handle it ad hoc, or not at all. They do just enough to get by.
+
+I’ve lived this firsthand. Most of my career has been in cybersecurity startups, where design is chronically scarce. The largest team I worked on in the last decade had fewer than ten designers at a company with over a thousand employees.
+
+So when Claude Design dropped recently and the usual online anxiety started up again, I had a different reaction than most. Not fear for my job, but optimism for a world with a higher floor of design quality.
+
+What tools like Claude Design represent isn’t a replacement for designers at businesses that already hire them. It’s design finally starting to reach the vast majority of businesses that never would have hired a designer in the first place.
+
+I don’t see that as a threat. I see it as a massive expansion of who gets to work with good design foundations. Designers included.
+
+**Appreciation isn’t craft**
+
+No tool automatically closes the gap between recognizing good design and being able to make it. It’s worth noting where most people actually stand in terms of recognizing good design in the first place: they don’t have particularly good taste (sorry, but it’s true). A smaller group has solid taste but doesn’t have the ability to execute on it. Those who can do both are a rare subset of an already rare group. (*If you’re reading this I imagine you fall into one of the two later camps*)
+
+There’s a [whole genre of video on YouTube](https://www.youtube.com/watch?v=n1JkFZIEY_k) built around this idea: give a novice top-of-the-line professional equipment, give an expert the most basic gear that gets the job done, and see what happens. Spoiler alert: the expert wins every time. Specific knowledge, built through years of reps, shapes the output in ways no tool can replicate.
+
+**What this means for you**
+
+If you’ve always cared about design but never had access to professional support, this is good news. The floor is rising. Solid design work that used to require a specialist is increasingly within reach.
+
+And if you’re a designer, the right response is confidence. Raising the floor doesn’t flatten your ceiling. If anything, the best designers stand to gain the most from these tools: more leverage, more reach, more impact than any previous generation has had access to.
+
+Think about the odds you’ve already beaten as a designer. You’re one of just a fraction of a percent of total workers and most companies don’t even practice your craft in any routine way. You chose a rare path, built real skills, and made it work despite the odds.
+
+New tools don’t change that. You’ll figure them out. Approach them with a beginner’s mind, staying curious and willing to evolve. That’s what every great craftsperson has always done: they find new ways to do their best work as new resources become available. This is no different.
+
+You beat the odds to get here. Now beat them again.
+
+Until next time,  
+Patrick
+
+---
+
+### 📚 Go deeper
+
+- **[Introducing Claude Design](https://www.anthropic.com/news/claude-design-anthropic-labs) — Anthropic** — The official company announcement.
+- **[“Be a Skill Surfer”](https://www.unknownarts.co/p/skill-surfer) — Unknown Arts** — The framing I keep coming back to as tools change faster than ever. Expertise is the thing worth protecting. Everything else is a wave to ride.
+- **[“Taste for Makers”](https://paulgraham.com/taste.html) — Paul Graham** — A short essay arguing that taste is real, learnable, and inseparable from doing great work. A useful counterweight to the “AI can do it now” narrative, because taste and craft are not the same thing and Graham explains why better than most.
