+++ src/pages/writing/index.astro
@@ -49,7 +49,7 @@ const posts = (await getCollection("writing", ({ data }) => !data.draft))
                 datetime={post.data.publishedDate.toISOString()}
                 class="mb-2 font-mono text-xs text-muted-foreground"
               >
-                {post.data.publishedDate.toLocaleDateString("en-US", { month: "short", day: "numeric", year: "numeric" })}
+                {post.data.publishedDate.toLocaleDateString("en-US", { timeZone: "UTC", month: "short", day: "numeric", year: "numeric" })}
               </time>
               <h2 class="mb-1 text-base font-semibold tracking-tight transition-colors group-hover:text-accent">
                 {post.data.title}
